Move SpectralStandard tests to new location.
